% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand.R
\name{logo}
\alias{logo}
\alias{logo_get}
\alias{logo_path}
\alias{logo_insert}
\title{Get Monash logo}
\source{
https://www.monash.edu/brandbook/brand-elements/our-logo
}
\usage{
logo_get(
  path = ".",
  stack = TRUE,
  color = c("blue", "black", "white"),
  type = c("png", "jpg", "ai"),
  hq = FALSE,
  overwrite = TRUE,
  filename = NULL
)

logo_path(
  stack = TRUE,
  color = c("blue", "black", "white"),
  type = c("png", "jpg", "ai"),
  hq = FALSE
)

logo_insert(
  path = NULL,
  stack = TRUE,
  color = c("blue", "black", "white"),
  type = c("png", "jpg", "ai"),
  hq = FALSE
)
}
\arguments{
\item{path}{the path to save the file to}

\item{stack}{TRUE for stacked logo, FALSE for one-line logo}

\item{color}{\code{blue}, \code{black} or \code{white}}

\item{type}{\code{png}, \code{jpg} or \code{ai}}

\item{hq}{TRUE for high-quality image (larger file size)}

\item{overwrite}{TRUE for overwriting (should it enquire?), FALSE for not. Not implemented yet.}

\item{filename}{A new file name for the logo. Not implemented yet.}
}
\description{
A quick way of getting the Monash logo.
}
\section{Functions}{
\itemize{
\item \code{logo_get}: Download the logo to provided path

\item \code{logo_path}: Get path to insert logo.

\item \code{logo_insert}: A wrapper to insert logo.
}}

