% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{logo_find}
\alias{logo_find}
\title{#' @rdname logo A wrapper to insert logo.
#' @export
logo_insert <- function(path = NULL,
stack = TRUE,
color = c("blue", "black", "white"),
type = c("png", "jpg", "ai"),
hq = FALSE) {
path <- path \%||\% rmarkdown::metadata$lib_dir \%||\% "."
knitr::include_graphics(logo_get(path = path,
stack = stack,
color = color,
type = type,
hq = hq,
overwrite = TRUE))
}}
\usage{
logo_find(logo_filename)
}
\description{
#' @rdname logo A wrapper to insert logo.
#' @export
logo_insert <- function(path = NULL,
stack = TRUE,
color = c("blue", "black", "white"),
type = c("png", "jpg", "ai"),
hq = FALSE) {
path <- path \%||\% rmarkdown::metadata$lib_dir \%||\% "."
knitr::include_graphics(logo_get(path = path,
stack = stack,
color = color,
type = type,
hq = hq,
overwrite = TRUE))
}
}
