% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monash-color.R
\name{monash-color}
\alias{monash-color}
\alias{color_all}
\alias{color_show}
\alias{color_primary}
\alias{color_secondary}
\title{Monash brand colors}
\usage{
color_all() # show both primary and secondary colors

color_show(print = TRUE)

color_primary()

color_secondary()
}
\arguments{
\item{print}{whether to print the color vector}
}
\description{
Monash brand colors
}
\section{Primary palette}{
\itemize{
\item Monash Blue is our most identifiable colour. It conveys youthfulness,
possibility and openness. It should always be considered first.
\item Black conveys prestige, timelessness and sophistication.
\item White is often shown as white space and conveys the brand personality
of being open and youthful.
\item Greys can be any percentage of black and provides additional flexibility
to our primary colour palette.
}

The primary colour palette is preferred for digital work.
}

\section{Secondary palette}{
Inspired by the colours of our academic robes, we have a range of bright,
colourful secondary colours.

Secondary colours are to be used:
\itemize{
\item in charts and diagrams to highlight key findings
\item as headings and subheadings
\item sparingly to provide highlights or accents – ideally one or two secondary
colours per double page spread
\item to speak to a particular audience group. For instance, colours that would
resonate best with our prospective undergraduate audience would be colourful,
bright and youthful. Industry and research would suit a more corporate/mature
choice to reflect focus and prestige.
}
}

