% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{quarto_template}
\alias{quarto_template}
\alias{quarto_template_use}
\alias{quarto_template_install}
\alias{quarto_template_add}
\title{List of Quarto templates available}
\usage{
quarto_template_use(
  type = c("report", "workingpaper", "thesis", "memo", "letter"),
  dir = type
)

quarto_template_install(
  type = c("report", "workingpaper", "thesis", "memo", "letter")
)

quarto_template_add(
  type = c("report", "workingpaper", "thesis", "memo", "letter")
)
}
\arguments{
\item{type}{One of either "report", "workingpaper", "thesis", "memo", or
"letter".}

\item{dir}{The name of the directory to put the template in. The directory
should not exist.}
}
\description{
This function lists the Monash Quarto templates available.
}
\examples{
\dontrun{
quarto_template_use("report", dir = "myreport")
quarto_template_install("workingpaper")
quarto_template_add("thesis")
}
}
