% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{release-functions}
\alias{release-functions}
\alias{release_lecture}
\alias{release_tutorial}
\alias{release_tutorial_solution}
\title{Release materials}
\usage{
release_lecture(
  week,
  dir = "tutorials",
  output_dir = "release",
  ignore = getOption("monash.lecture.ignore"),
  interactive = rlang::is_interactive(),
  overwrite = TRUE
)

release_tutorial(
  week,
  dir = "tutorials",
  output_dir = "release",
  ignore = getOption("monash.tutorial.ignore"),
  interactive = rlang::is_interactive(),
  overwrite = TRUE
)

release_tutorial_solution(
  week,
  dir = "tutorials",
  output_dir = "release",
  ignore = getOption("monash.tutorial.ignore"),
  interactive = rlang::is_interactive(),
  overwrite = TRUE
)
}
\arguments{
\item{week}{the week number}

\item{dir}{the subdirectory where the file is}

\item{output_dir}{the directory where the release folder is}

\item{ignore}{the file paths to ignore}

\item{interactive}{not used yet}

\item{overwrite}{should the files be overwritten?}
}
\description{
Release materials
}
\examples{
\dontrun{
release_lecture(9) # to release week 9 lecture
release_tutorial(9) # to release week 9 tutorial
release_tutorial_solution(9) # to release week 9 tutorial solution
}

}
