% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{working-dir}
\alias{working-dir}
\alias{create_bare_dir}
\alias{create_workshop_dir}
\alias{create_teaching_dir}
\title{Create skeleton working directories}
\usage{
create_bare_dir(
  code,
  name = code,
  full_name = name,
  destdir = code,
  layout = function() { },
  rstudio = rstudioapi::isAvailable(),
  open = interactive()
)

create_workshop_dir(
  name,
  full_name = name,
  destdir = getOption("monash.workshop_dir"),
  rstudio = TRUE,
  open = TRUE,
  layout = layout_workshop_dir
)

create_teaching_dir(
  code,
  name = code,
  full_name = name,
  destdir = getOption("monash.teaching_dir"),
  rstudio = TRUE,
  open = TRUE,
  layout = layout_teaching_dir
)
}
\arguments{
\item{code}{the code of the course}

\item{name}{the short name of the course. There should be no spaces.}

\item{full_name}{the full name of the course e.g. Wild-Caught Data}

\item{destdir}{the directory path}

\item{layout}{a function that creates the directory layout.
See \code{monash::layout_teaching_dir} for an example.}

\item{rstudio}{launch in rstudio?}

\item{open}{open after creating teaching directory?}
}
\description{
Functions to create a default directory structure.
}
\section{Functions}{
\itemize{
\item \code{create_bare_dir}: Create a directory structure specified by \code{layout}.

\item \code{create_workshop_dir}: Create a skeleton workshop directory.

\item \code{create_teaching_dir}: Create a skeleton teaching directory.
}}

