% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom.R
\name{zoom_attendance}
\alias{zoom_attendance}
\title{Link attendance record to google sheet}
\usage{
zoom_attendance(.data, .sheet, sheetname = "Lecture", week, upload = FALSE)
}
\arguments{
\item{.data}{The summary data frame with email, total and letter grade.}

\item{.sheet}{The link to the googlesheet}

\item{sheetname}{The name of the sheet}

\item{week}{The week number}
}
\description{
Link attendance record to google sheet
}
