% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom.R
\name{zoom_process}
\alias{zoom_process}
\title{Process the data frame from zoom meeting to total duration of attendance}
\usage{
zoom_process(
  .data,
  start = NA,
  end = NA,
  length = 120,
  accept = c(A = length * 1/2, P = length * 3/4)
)
}
\arguments{
\item{.data}{Data frame}

\item{start, end}{A date time of when the zoom meeting started or ended.
If NA, this is ignored. If the date time is supplied, then the time is censored.}

\item{length}{The total length of the session in minutes.}

\item{accept}{A named numeric vector that signifies the minimum required amount for the letter grade.}
}
\description{
Process the data frame from zoom meeting to total duration of attendance
}
