% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom.R
\name{zoom_read}
\alias{zoom_read}
\title{This reads in the zoom report}
\usage{
zoom_read(file, info = TRUE)
}
\arguments{
\item{file}{The file name for the zoom meeting report.}

\item{info}{TRUE or FALSE of whether the meeting information is included in the file.}
}
\description{
The zoom meeting report is expected to have the meeting
information and uncheck the "unique users" so that the
record for start and end time exists. The latter is
required so that total time for students is recorded
from the start of the lecture to the end of the time and
not because the student happens to be lingering before/after.
}
\details{
In order to identify the student, the zoom meeting should be
set to authenticate and restricted to Monash id alone.
This makes the data linkage easier with record in LMS.
}
