% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traveldiary.R
\name{create_simple_travel_diary}
\alias{create_simple_travel_diary}
\title{Create a simple travel diary for a Monash trip}
\usage{
create_simple_travel_diary(
  start_date,
  end_date,
  start_city = "Melbourne",
  destination,
  reason
)
}
\arguments{
\item{start_date}{The start date of the trip (as a Date object).}

\item{end_date}{The end date of the trip (as a Date object).}

\item{start_city}{The city where the trip starts (default is "Melbourne").}

\item{destination}{The destination city for the trip.}

\item{reason}{The reason for the trip (e.g., "Conference", "Meeting").}
}
\value{
A data frame representing the travel diary, with columns for Date, Location, and Description.
}
\description{
Creates a simple travel diary (consisting of one destination) for a trip. Used in conjunction with
\code{\link[=render_travel_diary]{render_travel_diary()}} to generate a PDF travel diary for uploading.
}
\examples{
create_simple_travel_diary(
  start_date = as.Date("2025-08-18"), 
  end_date = as.Date("2025-09-01"), 
  destination = "Perth", 
  reason = "NUMBATS Conference")

}
