% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monash-package.R
\docType{package}
\name{monash-package}
\alias{monash}
\alias{monash-package}
\title{monash: Consolidated tools and templates for staff at Monash University}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Rmarkdown templates and other tools for use at Monash University, Department of Econometrics & Business Statistics.
}
\author{
\strong{Maintainer}: Rob Hyndman \email{Rob.Hyndman@monash.edu} (\href{https://orcid.org/0000-0002-2140-5352}{ORCID})

Authors:
\itemize{
  \item Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID})
}

Other contributors:
\itemize{
  \item Mitchell O'Hara-Wild (\href{https://orcid.org/0000-0001-6729-7695}{ORCID}) [contributor]
}

}
\keyword{internal}
