% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monash-package.R
\docType{package}
\name{monash-package}
\alias{monash}
\alias{monash-package}
\title{monash: Consolidated tools and templates for staffs at Monash University}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID})

Authors:
\itemize{
  \item Rob Hyndman \email{Rob.Hyndman@monash.edu}
}

Other contributors:
\itemize{
  \item Mitchell O'Hara-Wild (\href{https://orcid.org/0000-0001-6729-7695}{ORCID}) [contributor]
}

}
\keyword{internal}
