% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{quarto_template}
\alias{quarto_template}
\alias{quarto_template_use}
\alias{quarto_template_install}
\alias{quarto_template_add}
\title{List of Quarto templates available}
\usage{
quarto_template_use(
  type = c("report", "workingpaper", "thesis", "memo", "letter"),
  dir = type
)

quarto_template_install(
  type = c("report", "workingpaper", "thesis", "memo", "letter")
)

quarto_template_add(
  type = c("report", "workingpaper", "thesis", "memo", "letter")
)
}
\description{
This function lists the Monash Quarto templates available.
}
