% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traveldiary.R
\name{render_travel_diary}
\alias{render_travel_diary}
\title{Render a travel diary to PDF}
\usage{
render_travel_diary(
  travel_diary,
  name,
  department = "Department of Econometrics and Business Statistics",
  output_path = getwd(),
  output_file,
  ...
)
}
\arguments{
\item{travel_diary}{A data frame representing the travel diary, created with \code{\link[=create_simple_travel_diary]{create_simple_travel_diary()}}.}

\item{name}{The name of the person for whom the travel diary is created.}

\item{department}{The department name (default is "Department of Econometrics and Business Statistics").}

\item{output_path}{The directory where the PDF will be saved (default is the current working directory).}

\item{output_file}{The name of the output PDF file (without extension).}

\item{...}{Additional arguments passed to \code{rmarkdown::render()}.}
}
\description{
Renders a travel diary to a PDF file using R Markdown. The travel diary must be created with \code{\link[=create_simple_travel_diary]{create_simple_travel_diary()}}.
}
\examples{
if (interactive()) {
diary <- create_simple_travel_diary(
   start_date = as.Date("2025-08-18"), 
   end_date = as.Date("2025-09-01"), 
   destination = "Perth", 
   reason = "NUMBATS Conference") |> 
render_travel_diary(output_file = "test.pdf", name = "Michael")
}

}
