% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{working-dir}
\alias{working-dir}
\alias{create_layout_dir}
\alias{create_workshop_dir}
\alias{create_teaching_dir}
\title{Create skeleton working directories}
\usage{
create_layout_dir(
  code,
  name = code,
  proj_name = paste0(name, "-private"),
  full_name = name,
  destdir = code,
  layout = NULL,
  rstudio = rstudioapi::isAvailable(),
  open = interactive()
)

create_workshop_dir(
  code,
  name = code,
  full_name = name,
  destdir = getOption("monash.workshop.dir"),
  rstudio = TRUE,
  open = TRUE
)

create_teaching_dir(
  code,
  name = code,
  full_name = name,
  destdir = getOption("monash.teaching.dir"),
  rstudio = TRUE,
  open = TRUE
)
}
\arguments{
\item{code}{the code of the course}

\item{name}{the short name of the course. There should be no spaces.}

\item{proj_name}{the project name for Rpro}

\item{full_name}{the full name of the course e.g. Wild-Caught Data}

\item{destdir}{the directory path}

\item{layout}{the yaml file name (TODO: or a list)}

\item{rstudio}{launch in rstudio?}

\item{open}{open after creating teaching directory?}
}
\description{
Functions to create a default directory structure.
}
\section{Functions}{
\itemize{
\item \code{create_layout_dir}: Create a directory structure specified by \code{layout}.

\item \code{create_workshop_dir}: Create a skeleton workshop directory.

\item \code{create_teaching_dir}: Create a skeleton teaching directory.
}}

